The offset of a field is within an int32.
git-svn-id: https://llvm.org/svn/llvm-project/vmkit/trunk@60247 91177308-0d34-0410-b5e6-96231b3b80d8

+++ lib/JnJVM/VMCore/JavaClass.h
@@ -1136,7 +1136,7 @@ public:
   /// ptrOffset - The offset of the field when the object containing
   /// the field is casted to an array of bytes.
   ///
-  uint64 ptrOffset;
+  uint32 ptrOffset;
   
   /// num - The index of the field in the field list.
   ///
